Add the exercise 2.1 from "The C programming language"
Prints the ranges of the different types. This implementation
depends on the length of the steps used, maybe they are too big
for a machine with less power. The ranges are directly computed
by adding or subtracting values until a shift in the value.
Of course there are better solutions to this.